requirements.txt
streamlit>=1.38.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
crewai>=0.30.0
